requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
cryptography==41.0.7
paramiko==3.4.0
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
anthropic>=0.40.0
google-generativeai>=0.8.0
